indra-testlib: Break out testing utils into separate project
This project is not published, and helps with IDE integration.
