+++ indra-common/src/test/java/net/kyori/indra/util/VersioningTest.java
@@ -28,8 +28,8 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.stream.Stream;
-import net.kyori.indra.IndraTesting;
 import net.kyori.indra.git.GitPlugin;
+import net.kyori.indra.test.IndraTesting;
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.errors.GitAPIException;
 import org.gradle.api.JavaVersion;
