+++ indra-testlib/gradle.properties
@@ -1,0 +1,1 @@
+indra.skipPublishing=true
