+++ indra-testlib/src/main/java/net/kyori/indra/test/IndraFunctionalTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.indra;
+package net.kyori.indra.test;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
